Functional pong game with no UI or sound
Game now starts with 1 sec delay with the ball at the center and a
random velocity direction.

When the ball goes out of bounds by colliding with one of the goals, the
game restarts.

Game keeps score of any goals.

Game can be restarted by pressing the `R` key on a QWERTY keyboard while the
ball is in play. Else, it does nothing.